+++ src/Pages/AboutUs/Certificates/Certificates.module.scss
@@ -6,19 +6,13 @@
     width: 100%;
 }
 
-h1 {
-    margin-top: 63px;
-    color: white;
-    max-width: 200px;
-    width: 100%;
-}
-
 .containerCards {
     justify-content: center;
     display: flex;
     gap: 20px;
     max-width: 100%;
     flex-wrap: wrap;
+    padding-bottom: 130px;
 }
 
 .certificado {
@@ -48,7 +42,8 @@ h2 {
 
 .flex {
     width: 100%;
-    margin-left: 30px;
+    margin-left: 40px;
+    margin-top: 20px;
 }
 
 .img {
