+++ src/Pages/AboutUs/Certificates/index.ts
@@ -1,0 +1,1 @@
+export * from './Certificates'
