+++ src/Pages/Home/Components/CTATourVirtual/Assets/Images/index.ts
@@ -1,3 +1,3 @@
 export {default as icon360} from './icon360.svg'
 export {default as iconPlay} from './iconPlay.svg'
-export {default as schoolImage} from './schoolImage.png'
+export {default as schoolImage} from './schoolImage.jpeg'
